fix: exercise 06 of TOP
Modify style.css. Move class selector .small-para so it is applied
after .para. The paragraph is now smaller than the above paragraph.
Move class selector .confirm after .button so that the green background
is applied at last. Add .text selector to .child. It now matches the
specificity of div .text and applies a smaller text for the last div.